предрелизная версия(отредактированная версия)
