+++ src/main/java/com/electronicsstore/store/controllers/UserController.java
@@ -4,7 +4,6 @@ import com.electronicsstore.store.models.User;
 import com.electronicsstore.store.repo.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
