+++ src/main/java/com/electronicsstore/store/repo/ProductRepository.java
@@ -1,7 +1,6 @@
 package com.electronicsstore.store.repo;
 
 import com.electronicsstore.store.models.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ProductRepository extends CrudRepository<Product, Long> {
