+++ src/main/resources/templates/blocks/cart.html
@@ -1,7 +1,5 @@
 <!-- Карточка товара -->
 
-
-
 <div class="cart-item" th:fragment="card">
   <div class="item-info">
     <a th:href="'/about-product/' + ${el.product_id}">
